Make illegal access warning go away.
We MUST modify the unmodifiable map that contains the environment variables.
Otherwise the extension cannot do it's thing. This commit
adds the necessary tweaks to the Maven Failsafe JVM command line args to
open up the module for this. (It's not enough to go for MAVEN_OPTS etc,
as the failsafe plugin starts a new JVM!!!)

See also payara/ecosystem-support#30